I implement the first version of the CI/CD functionality where the user can deploy a Jemo plugin from a Github repository. Currently, only public repositories are supported.

+++ core/src/main/java/org/eclipse/jemo/sys/internal/Util.java
@@ -249,7 +249,11 @@ public class Util {
     }
 
     public static final boolean deleteDirectory(File dir) {
-        Arrays.asList(dir.listFiles()).stream().forEach(f -> {
+        if (!Files.exists(dir.toPath())) {
+            return false;
+        }
+
+        Arrays.asList(dir.listFiles()).forEach(f -> {
             if (f.isDirectory()) {
                 deleteDirectory(f);
             } else {
